@@ -3,22 +3,23 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@babel/parser": "^7.15.7",
-		"@material-ui/core": "^4.12.3",
-		"@material-ui/lab": "^4.0.0-alpha.60",
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^12.1.0",
-		"@testing-library/user-event": "^13.2.1",
-		"@uiw/react-md-editor": "^3.6.3",
-		"axios": "^0.21.4",
+		"@babel/parser": "^7.19.1",
+		"@material-ui/core": "^4.12.4",
+		"@material-ui/lab": "^4.0.0-alpha.61",
+		"@testing-library/jest-dom": "^5.16.5",
+		"@testing-library/react": "^13.4.0",
+		"@testing-library/user-event": "^14.4.3",
+		"@uiw/react-md-editor": "^3.18.1",
+		"assert": "^2.0.0",
+		"axios": "^0.27.2",
 		"babel-plugin-transform-react-remove-prop-types": "^0.4.24",
 		"balloon-css": "^1.2.0",
-		"prop-types": "^15.7.2",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"prop-types": "^15.8.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-loading": "^2.0.3",
-		"react-router-dom": "^5.3.0",
-		"react-scripts": "4.0.3"
+		"react-router-dom": "5.3.0",
+		"react-scripts": "^5.0.1"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -43,7 +44,8 @@
 	},
 	"devDependencies": {
 		"babel-eslint": "^10.1.0",
-		"eslint": "^7.32.0",
-		"eslint-plugin-react": "^7.26.0"
+		"eslint": "^8.23.1",
+		"eslint-plugin-react": "^7.31.8",
+		"yarn-upgrade-all": "^0.7.1"
 	}
 }
